Init branch and add net_address class
There are many features of bitcoin's net_address that we don't need now,
so I wrote it myself. And now the net address only support IPv4. Later
when we start to manage addresses I will add more kinds of addresses.

It seems that std didn't offer us a good lib to convert ip string to
byte array, so I just implemented a simple function which is not robust,
I will enhance it later.