Port bitcoin serialization
In order to make it compatible with all bitcoin types and structres the decision was made
that it's better to keep serialization as is and adjust our code to their vision.
Though there's almost no difference as you can see, later it will be easier for everyone
to follow the prescription if we keep it this way.