Added stream-like vector class VStream
This is copied from the CDataStream class in stream.h in
bitcoin project but eliminated most of the functions that
we don't use. Basically, it uses a vector of char with
customized allocator (zero_after_free_allocator), and
added stream-like behaviors on top of it (e.g., read, write,
operator<< operator>>). The zero_after_free_allocator
guarentees that the contents are cleared before
the vector is deallocated.

*Note: in the original bitcoin project, zeroafterfree.h
includes support/cleanse.h for clearing contents in the deallocate
function, basically to support Windows machines. Since we only
support Unix, we remove this dependency and simply replace the
function memory_cleanse with memset_s.

This commit also modifies the TestSer in which the usage of
std::stringstream is replaced with this VSteam. There is also
other minor changes such as modifiying .gitignore and reformating code.