Added stream-like vector class VStream (#18)
Stream-like vector class VStream

This is mainly copied from the CDataStream class in stream.h in
bitcoin project but eliminated most of the functions that
we don't use. Basically, it uses a vector of char with
customized allocator (zero_after_free_allocator), and
added stream-like behaviors on top of it (e.g., read, write,
operator<< operator>>). The zero_after_free_allocator
guarentees that the contents are cleared before
the vector is deallocated.

This commit also modifies the TestSer class in which the usage of
std::stringstream is replaced with this VSteam. hash.h is also modified
in the same manner. 

There is also other minor changes such as modifiying .gitignore and reformating code.