Add VERIFY opcode to tasm
Create test cases for verification of transactions' input and output scripts
The usage of tasm in phase 1 is covered in test cases
Modify the structure of internal tasm execution, external API is changed a bit.
To achive a termination of a program which is comprised of subprograms or op codes in most of cases at least for now we insert FAIL opcode after each subprogram.
Preprocessor is modified in a way that it's adjusted to the modified execution ABI and inserts FAIL and SUCCESS opcodes at the end or the program. As before the
need of preproccesor is to reduce the amount of data sent over the network.