Add deserialization to vstream
In order to use vstream as the universal thing with
a property or being a vector and a stream at the same time
we need to be able to serialize and deserialize both to the external streams
and to the vstream itself. Before this commit the serialization vstream to vstream
did a concatination instead. This behavior is removed. Deserialization and serialization
are done in the way that classes in this project use via macro and serializationOp helper function.
The creation of the method is delegated to the serialization library (serialize.h)