Change codestyle of tasm
Rename functors and tasm to .h files. Change name of the methods and
classes to the google style we're currently using. Rebase to the latest master.
Fix some issues with GetOptimalEncoding after rebase.
Completely remove script from the project.
Add necessary includes for the linux machine. It was tested on pascal.
Address comments from PR.