Fix the multi-thread solve bug
When one of the threads in the threadpool successfully finds a valid
nonce, the main thread goes on while the other threads might not have
stopped. This may cause a runtime error when the main thread has
finished the Solve method and final_nonce goes out of scope, but some
other thread is still trying accessing it.