Add header only implementation of the CRC32C algorithm
I rewrote the crc32 method with a different polynomial (0x1EDC6F41).
Changing the polynomial allowed me to take advantage of the crc32
instructions present in the SSE42 extension. Furhter using the
PCLMULQDQ instructions I was able to pipeline the crc32 instruction
to achieve even greater efficiency.

The overall speedup with all accelerations enabled is around 60x.
But since the underlying hardware might not support those I also
implemented a version that is hardware independent.

The final result of the crc32c algorithm is negated to comply with the
standard implementation.